Trait proc_macro_roids::DeriveInputStructExt

pub trait DeriveInputStructExt {
    fn data_struct(&self) -> &DataStruct;

    fn data_struct_mut(&mut self) -> &mut DataStruct;

    fn fields(&self) -> &Fields;

    fn fields_mut(&mut self) -> &mut Fields;

    fn fields_named(&self) -> &FieldsNamed;

    fn fields_named_mut(&mut self) -> &mut FieldsNamed;

    fn is_unit(&self) -> bool;

    fn is_named(&self) -> bool;

    fn is_tuple(&self) -> bool;

    fn assert_fields_unit(&self);

    fn assert_fields_named(&self);

    fn assert_fields_unnamed(&self);
}

Functions to make it ergonomic to work with struct ASTs.

Required methods

fn data_struct(&self) -> &DataStruct

Returns a reference to the data_struct of a struct's AST.

Panics

Panics if the AST is not for a struct.

fn data_struct_mut(&mut self) -> &mut DataStruct

Returns a mutable reference to the data_struct of a struct's AST.

Panics

Panics if the AST is not for a struct.

fn fields(&self) -> &Fields

Returns a reference to the fields of a struct's AST.

Panics

Panics if the AST is not for a struct.

fn fields_mut(&mut self) -> &mut Fields

Returns a mutable reference to the fields of a struct's AST.

Panics

Panics if the AST is not for a struct.

fn fields_named(&self) -> &FieldsNamed

Returns a mutable reference to the named fields of a struct's AST.

Panics

Panics if the AST is not for a struct with named fields.

fn fields_named_mut(&mut self) -> &mut FieldsNamed

Returns a mutable reference to the named fields of a struct's AST.

Panics

Panics if the AST is not for a struct with named fields.

fn is_unit(&self) -> bool

Returns true if the AST is for a unit struct.

fn is_named(&self) -> bool

Returns true if the AST is for a struct with named fields.

fn is_tuple(&self) -> bool

Returns true if the AST is for a struct with unnamed fields.

fn assert_fields_unit(&self)

Panics if the AST is not for a unit struct.

fn assert_fields_named(&self)

Panics if the AST is not for a struct with named fields.

fn assert_fields_unnamed(&self)

Panics if the AST is not for a struct with unnamed fields.

Loading content...