Trait proc_macro_roids::FieldsExt

pub trait FieldsExt {
    fn is_unit(&self) -> bool;

    fn is_named(&self) -> bool;

    fn is_tuple(&self) -> bool;

    fn construction_form(&self) -> TokenStream;
}

Functions to make it ergonomic to work with Fields.

Required methods

fn is_unit(&self) -> bool

Returns true if the Fields is for a unit struct.

fn is_named(&self) -> bool

Returns true if the Fields is for a struct with named fields.

fn is_tuple(&self) -> bool

Returns true if the Fields is for a struct with unnamed fields.

fn construction_form(&self) -> TokenStream

Returns a token stream of the construction form of the fields.

For unit fields, this returns an empty token stream.

Tuple fields: (_0, _1,)

Named fields: { field_0, field_1 }

Examples

Loading content...